Removed Server Side Rendering
