+++ website/pages/login/index.tsx
@@ -8,7 +8,6 @@ import Input from "@material-ui/core/Input";
 import InputLabel from "@material-ui/core/InputLabel";
 import Paper from "@material-ui/core/Paper";
 import useInputState from "../../hooks/useInputState";
-import { GetServerSideProps } from "next";
 import { Facebook } from "@material-ui/icons";
 
 const styles = createStyles((theme: Theme) => ({
@@ -139,6 +138,13 @@ function LoginPage(props: Props) {
     toggleLogin(!isLogin);
   };
 
+  React.useEffect(() => {
+    let status = getParameterByName("status")
+    if (status == "success") {
+      window.location.href = "/"
+    }
+  }, [])
+
   return (
     <main className={classes.main}>
       {isLogin ? (
@@ -222,112 +228,108 @@ function LoginPage(props: Props) {
           </Button>
         </Paper>
       ) : (
-        <Paper className={classes.paper}>
-          <Avatar
-            className={classes.avatar}
-            src="https://img.icons8.com/officel/80/000000/court-judge.png"
-          />
-          <Typography variant="h3" gutterBottom>
-            Register
+          <Paper className={classes.paper}>
+            <Avatar
+              className={classes.avatar}
+              src="https://img.icons8.com/officel/80/000000/court-judge.png"
+            />
+            <Typography variant="h3" gutterBottom>
+              Register
           </Typography>
-          <form
-            className={classes.form}
-            onSubmit={(e) => {
-              e.preventDefault();
-              reset();
-            }}
-          >
-            <FormControl margin="normal" required fullWidth>
-              <InputLabel htmlFor="username">Username</InputLabel>
-              <Input
-                value={value.username}
-                onChange={handleChange}
-                id="username"
-                name="username"
-                autoFocus
-              />
-            </FormControl>
-            <FormControl margin="normal" required fullWidth>
-              <InputLabel htmlFor="email">Email</InputLabel>
-              <Input
-                value={value.email}
-                onChange={handleChange}
-                id="email"
-                name="email"
-                autoFocus
-              />
-            </FormControl>
-            <FormControl margin="normal" required fullWidth>
-              <InputLabel htmlFor="password">Password</InputLabel>
-              <Input
-                value={value.password}
-                onChange={handleChange}
-                id="password"
-                name="password"
-                autoFocus
-              />
-            </FormControl>
-            <Button
-              variant="contained"
-              type="submit"
-              fullWidth
-              color="primary"
-              className={classes.submit}
+            <form
+              className={classes.form}
+              onSubmit={(e) => {
+                e.preventDefault();
+                reset();
+              }}
             >
-              SignUp
+              <FormControl margin="normal" required fullWidth>
+                <InputLabel htmlFor="username">Username</InputLabel>
+                <Input
+                  value={value.username}
+                  onChange={handleChange}
+                  id="username"
+                  name="username"
+                  autoFocus
+                />
+              </FormControl>
+              <FormControl margin="normal" required fullWidth>
+                <InputLabel htmlFor="email">Email</InputLabel>
+                <Input
+                  value={value.email}
+                  onChange={handleChange}
+                  id="email"
+                  name="email"
+                  autoFocus
+                />
+              </FormControl>
+              <FormControl margin="normal" required fullWidth>
+                <InputLabel htmlFor="password">Password</InputLabel>
+                <Input
+                  value={value.password}
+                  onChange={handleChange}
+                  id="password"
+                  name="password"
+                  autoFocus
+                />
+              </FormControl>
+              <Button
+                variant="contained"
+                type="submit"
+                fullWidth
+                color="primary"
+                className={classes.submit}
+              >
+                SignUp
             </Button>
-            <Typography color="error"></Typography>
-            <Button
-              variant="outlined"
-              className={classes.googleButton}
-              size="large"
-              fullWidth
-              onClick={() => googleLogin()}
-            >
-              <img
-                src="https://img.icons8.com/color/24/000000/google-logo.png"
-                className={classes.signInIcon}
-              />
+              <Typography color="error"></Typography>
+              <Button
+                variant="outlined"
+                className={classes.googleButton}
+                size="large"
+                fullWidth
+                onClick={() => googleLogin()}
+              >
+                <img
+                  src="https://img.icons8.com/color/24/000000/google-logo.png"
+                  className={classes.signInIcon}
+                />
               Sign Up with Google
             </Button>
+              <Button
+                variant="outlined"
+                color="secondary"
+                size="large"
+                fullWidth
+                onClick={() => facebookLogin()}
+                className={classes.facebookButton}
+              >
+                <Facebook className={classes.signInIcon} />
+              Sign Up with Facebook
+            </Button>
+            </form>
             <Button
               variant="outlined"
               color="secondary"
-              size="large"
-              fullWidth
-              onClick={() => facebookLogin()}
-              className={classes.facebookButton}
+              onClick={() => toggleisLogin()}
+              style={{ border: "None", outline: "none" }}
             >
-              <Facebook className={classes.signInIcon} />
-              Sign Up with Facebook
-            </Button>
-          </form>
-          <Button
-            variant="outlined"
-            color="secondary"
-            onClick={() => toggleisLogin()}
-            style={{ border: "None", outline: "none" }}
-          >
-            Already Registered? Login here.
+              Already Registered? Login here.
           </Button>
-        </Paper>
-      )}
+          </Paper>
+        )}
     </main>
   );
 }
 
-export const getServerSideProps: GetServerSideProps = async (context) => {
-  var data = context.query;
-  if (data.status) {
-    if (data.status == "success") {
-      context.res.statusCode = 302;
-      context.res.setHeader("Location", "/");
-    }
-  }
-
-  return {
-    props: {},
-  };
-};
+function getParameterByName(name, url = window.location.href) {
+  if (!url) url = window.location.href;
+  name = name.replace(/[\[\]]/g, '\\$&');
+  var regex = new RegExp('[?&]' + name + '(=([^&#]*)|&|#|$)'),
+    results = regex.exec(url);
+  if (!results) return null;
+  if (!results[2]) return '';
+  return decodeURIComponent(results[2].replace(/\+/g, ' '));
+}
 
 export default withStyles(styles)(LoginPage);
